% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTime.R
\name{dbTime}
\alias{dbTime}
\title{Time Information about Tables and Views}
\usage{
dbTime(channel)
}
\arguments{
\item{channel}{an RODBC connection.}
}
\value{
Data frame containing four columns:
\item{name}{name of table/view.}
\item{type}{type of table/view.}
\item{create_date}{date created.}
\item{modify_date}{date last modified.}
}
\description{
Get time information about tables and views: when they were created and when
they were last modified.
}
\note{
The date and and time information is queried from \verb{sys.tables} and
\verb{sys.views}. If these do not exist, the function will return a harmless
ODBC error.
}
\examples{
\dontrun{
con <- odbcConnect("myDatabase")

dbTime(con)
}

}
\seealso{
\code{\link{Sys.time}} is the base function to show the current date and
time.

\code{\link{dbTools-package}} gives an overview of the package.
}
