% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools-package.R
\docType{package}
\name{dbTools-package}
\alias{dbTools-package}
\alias{dbTools}
\title{Tools to Work with RODBC Databases}
\description{
Simple tools that extend the functionality of the RODBC package.
}
\details{
\emph{Browse database:}
\tabular{ll}{
  \code{\link{dbOverview}} \tab Dimensions and column names\cr
  \code{\link{dbStorage}}  \tab Storage size\cr
  \code{\link{dbTime}}     \tab Time created and modified
}
\emph{Browse table:}
\tabular{ll}{
  \code{\link{tableDim}}      \tab Dimensions\cr
  \code{\link{tableHead}}     \tab First rows\cr
  \code{\link{tableNcol}}     \tab Number of columns\cr
  \code{\link{tableNrow}}     \tab Number of rows\cr
  \code{\link{tableOverview}} \tab Data types and dimensions
}
\emph{Helper functions:}
\tabular{ll}{
  \code{\link{tableQuote}} \tab Quote table name
}
}
\author{
Arni Magnusson.
}
