% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RODBCtools-package.R
\docType{package}
\name{RODBCtools-package}
\alias{RODBCtools-package}
\title{Tools to Work with RODBC Databases}
\description{
Simple tools that extend the functionality of the RODBC package.
}
\details{
\emph{Table:}
\tabular{ll}{
  \code{\link{tableNcol}} \tab Number of columns\cr
  \code{\link{tableNrow}} \tab Number of rows
}
}
\author{
Arni Magnusson.
}
