% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableDim.R
\name{tableDim}
\alias{tableDim}
\title{Table Dimensions}
\usage{
tableDim(channel, sqtable)
}
\arguments{
\item{channel}{an RODBC connection.}

\item{sqtable}{a database table or view.}
}
\value{
Vector of length two, containing the number of rows and columns.
}
\description{
Return the number of rows and columns in a database table.
}
\examples{
\dontrun{
con <- odbcConnect("myDatabase")

tableDim(con, "sysusers")
}

}
\seealso{
\code{\link{tableNrow}} and \code{\link{tableNcol}} are the underlying
functions to get the number of rows and columns in a database table.

\code{\link{dim}} is the base function to return the dimensions for data
frames inside the R workspace.

\code{\link{MSSQL-package}} gives an overview of the package.
}
