% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNcol.R
\name{tableNcol}
\alias{tableNcol}
\title{Number of Columns}
\usage{
tableNcol(channel, sqtable)
}
\arguments{
\item{channel}{an RODBC connection.}

\item{sqtable}{a database table or view.}
}
\value{
Number of columns as integer.
}
\description{
Return the number of columns in a database table.
}
\examples{
\dontrun{
con <- odbcConnect("myDatabase")

tableNcol(con, "sysusers")
}

}
\seealso{
\code{\link{tableDim}} and \code{\link{tableNrow}} also return the dimensions
of a database table.

\code{\link[RODBC]{sqlColumns}} is the underlying function used to examine
the table columns.

\code{\link{ncol}} is the base function to return the number of columns for
data frames inside the R workspace.

\code{\link{MSSQL-package}} gives an overview of the package.
}
