% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableQuote.R
\name{tableQuote}
\alias{tableQuote}
\title{Quote Table Name}
\usage{
tableQuote(sqtable)
}
\arguments{
\item{sqtable}{table name, with or without schema name.}
}
\value{
String with special quotes.
}
\description{
Add special quotes around table name.
}
\note{
The \code{sqlQuery} function requires special quotes if the table name has
spaces. Furthermore, the schema name must not be inside the special quotes.
}
\examples{
tableQuote("table")
tableQuote("table name")
tableQuote("schema.table")
tableQuote("schema.table name")

}
\seealso{
\code{\link[RODBC]{sqlQuery}} requires special quotes if the table name has
spaces.

\code{\link{Quotes}} in base R.

\code{\link{MSSQL-package}} gives an overview of the package.
}
