% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDate.R
\name{dbDate}
\alias{dbDate}
\title{Date Information about Tables and Views}
\usage{
dbDate(channel)
}
\arguments{
\item{channel}{an RODBC connection.}
}
\value{
Data frame containing four columns:
\item{name}{name of table/view.}
\item{type}{type of table/view.}
\item{create_date}{date created.}
\item{modify_date}{date last modified.}
}
\description{
Get date and time information about tables and views: when they were created
and when they were last modified.
}
\examples{
\dontrun{
dbDate(con)
}

}
\seealso{
\code{\link{Sys.time}} is the base function to show the current date and
time.

\code{\link{RODBCtools-package}} gives an overview of the package.
}
