% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools-package.R
\docType{package}
\name{dbTools-package}
\alias{dbTools-package}
\alias{dbTools}
\title{Tools to Work with RODBC Databases}
\description{
Simple tools that extend the functionality of the RODBC package.
}
\details{
\emph{Database:}
\tabular{ll}{
  \code{\link{dbDate}}  \tab Date information
}
\emph{Table:}
\tabular{ll}{
  \code{\link{tableDim}}  \tab Dimensions\cr
  \code{\link{tableNcol}} \tab Number of columns\cr
  \code{\link{tableNrow}} \tab Number of rows
}
}
\author{
Arni Magnusson.
}
